requirements.txt
bleach
boto
celery
django
django-allauth
django-celery-email
django-cleanup
django-crispy-forms
django-filter
django-resized
django-simple-captcha
djangorestframework
django-storages-redux
dj-database-url
markdown
mysqlclient
newrelic
psycopg2
sendgrid-django
tweepy
waitress
